+++ src/easyHttp.h
@@ -104,6 +104,6 @@ static size_t write_filedata(void *ptr, size_t size, size_t nmemb, void *stream)
       
 static size_t write_filedata(void *ptr, size_t size, size_t nmemb, void *stream)
 {
-	int written = fwrite(ptr, size, nmemb, (FILE *)stream);
+	size_t written = fwrite(ptr, size, nmemb, (FILE *)stream);
     return written;
 }
